# server/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Message:
    properties:
      content:
        type: string
      id:
        type: integer
      room:
        $ref: '#/definitions/model.Room'
      roomId:
        type: string
      sender:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Associations
      senderId:
        type: integer
      sentAt:
        type: string
    type: object
...
  model.Room:
    properties:
      attendees:
        items:
          $ref: '#/definitions/model.User'
        type: array
      consolidated:
        type: string
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      host:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Associations
      hostId:
        type: integer
      id:
        type: string
      imageUrl:
        type: string
      isClosed:
        type: boolean
      isPrivate:
        type: boolean
      name:
        type: string
      noOfAttendees:
        type: integer
      time:
        type: string
      updatedAt:
        type: string
      venue:
        type: string
      venuePlaceId:
        type: string
      venueUrl:
        type: string
    type: object
...
  model.User:
    properties:
      email:
        type: string
      friends:
        items:
          $ref: '#/definitions/model.User'
        type: array
      id:
        type: integer
      isEmailValid:
        type: boolean
      isOnline:
        type: boolean
      lastSeen:
        type: string
      noOfFriends:
        type: integer
      noOfPendingFriendRequests:
        type: integer
      noOfRoomInvites:
        type: integer
      noOfRooms:
        type: integer
      password:
        type: string
      pictureUrl:
        type: string
      registeredAt:
        type: string
      rooms:
        description: Associations
        items:
          $ref: '#/definitions/model.Room'
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  request.CreateMessageRequest:
    properties:
      content:
        type: string
    type: object
...
  request.GoogleAuthRequest:
    properties:
      code:
        type: string
    type: object
...
  request.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  request.ResetPasswordRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  request.SendOTPEmailRequest:
    properties:
      email:
        type: string
      purpose:
        type: string
    type: object
...
  request.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    type: object
...
  response.AuthResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      pictureUrl:
        type: string
      username:
        type: string
    type: object
...
  utils.EmptyApiResponse:
    properties:
      data:
        type: object
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for JustJio.
  title: JustJio API
  version: "1.0"
paths:
  /auth/google:
    post:
      consumes:
      - application/json
      description: Authenticates a user using Google OAuth authorization code
      parameters:
      - description: Google OAuth request
        in: body
        name: google
        required: true
        schema:
          $ref: '#/definitions/request.GoogleAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated via Google successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/response.AuthResponse'
              message:
                type: string
              status:
                type: string
              token:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Google OAuth login
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with username and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/response.AuthResponse'
              message:
                type: string
              status:
                type: string
              token:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Login
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password after OTP verification
      parameters:
      - description: Password reset request
        in: body
        name: reset
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Reset password
      tags:
      - Authentication
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: Sends a One-Time Password to the user's email for verification
        or password reset
      parameters:
      - description: OTP email request
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/request.SendOTPEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "400":
          description: Invalid input or invalid purpose
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: Email already verified
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Send OTP email
      tags:
      - Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email verification
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: User signed up successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/response.AuthResponse'
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Sign up
      tags:
      - Authentication
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies the One-Time Password sent to the user's email
      parameters:
      - description: OTP verification request
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/request.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "400":
          description: Invalid input or invalid OTP
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: User not found or OTP not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Verify OTP
      tags:
      - Authentication
...
  /messages/{msgId}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific message by its ID
      parameters:
      - description: Message ID
        in: path
        name: msgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved message successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/model.Message'
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: No message found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Get message by ID
      tags:
      - Messages
...
  /rooms/{roomId}/messages:
    get:
      consumes:
      - application/json
      description: Retrieves paginated messages for a specific room with optional
        sorting
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: true
        description: Sort order (ascending if true)
        in: query
        name: asc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved messages successfully
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.Message'
                type: array
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: No messages found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Get room messages
      tags:
      - Messages
    post:
      consumes:
      - application/json
      description: Creates and saves a new message in the specified room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/request.CreateMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Message saved successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Room or user not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Create message
      tags:
      - Messages
...
schemes:
- http
- https
swagger: "2.0"